patch(`postLbp`): Migrated post deployment tasks between phase 1 and 2 of post LBP deploy
